refactor(StationAcceptanceTest): 테스트 코드 리팩토링
- fixture 분리 (setUp, method)
- 가독성을 위해 메소드명, 변수명 수정
