+++ src/test/java/nextstep/subway/station/StationAcceptanceTestFixture.java
@@ -1,0 +1,53 @@
+package nextstep.subway.station;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import java.util.HashMap;
+import java.util.Map;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.http.MediaType;
+
+class StationAcceptanceTestFixture {
+
+    @LocalServerPort
+    int port;
+
+    @BeforeEach
+    public void setUp() {
+        RestAssured.port = port;
+    }
+
+    /**
+     * 지하철역 조회
+     *
+     * @return 지하철역 조회 Response
+     */
+    protected static ExtractableResponse<Response> 지하철역_조회() {
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().get("/stations")
+                .then().log().all()
+                .extract();
+    }
+
+    /**
+     * 지하철역 생성
+     *
+     * @param stationName 생성할 지하철역 이름
+     * @return 생성된 지하철역 id
+     */
+    protected static ExtractableResponse<Response> 지하철역_생성(String stationName) {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", stationName);
+
+        return RestAssured.given().log().all()
+                .body(params)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().post("/stations")
+                .then().log().all()
+                .extract();
+    }
+
+}
